Initial DatabaseSystem
This commit adds the data structures which the Database System primarily
uses for it's operation. The data structures handle both the program's
storage of the database and SQL queries to said database. Methods are
also in place for manipulation of the data from the database, though
they are not fully tested.

A demo which shows the queries the structures uses along with the
addition and removal of question. The demo will create a database name
TestDatabase.db if it doesn't exist and add to it if there are less than
5 questions. It will remove questions if there are 5 or more (note that
this is set up in the demo and the system itself).