Database Simpleton
The class QuestionSource creates a simpleton object which can contain
Databases and pull questions from them randomly and pull a previously
used question.

There are two constructors available for the class, an empty constructor
and a constructor which requires a String. The key difference is that
the String constructor which use the String to select a folder and load
database files (ending with .db or .sqlite) from the folder. You can add
more database by adding a database object , addDatabase(Database), or
giving it another folder to load from , addDatabasesFrom(String). Once
databases are set up, you can pull a random question from the source,
randomQuestion(), or pull an old question, this[int]. You can get the
number of questions from QuestionCount and clear the source of question
allowing you to get used questions randomly again.

You can use DatabaseDriver solution to see the source in action.