Database Manipulation
The database manipulation system is complete and has been though basic
testing. Table names are not protected from SQL Infection, but other
than that there are no known bugs that impact functionally. Setting up
the system should be a simple matter as it just requires to add the
DBManipulate control to a Form. If the control isn't available in the
toolbox, you can also create the control in the Form's cs file and add
it to the form with form.Controls.Add(vName).

The system is GUI-based and shouldn't be hard to figure out. I'll avoid
stating how to use the GUI since I want feedback on how easy is it to
figure out and if I need to redesign it based on feedback.