Source Save and Loading
This commit include updates that can save and load QuestionSource. Some
bugs in DBManipulate are also fixed (namely Dropped tables not
disappearing when they should).

To save a QuestionSource, just use the toSave() method in QuestionSource
and that the returned String.

To load a QuestionSource, take the same String returned from toSave()
and use it in the constructor which supports a single String. If any
databases have been changed since the Source has been saved, then the
constructor will throw a InvalidDataException.